+++ res/layout/login_activity.xml
@@ -75,13 +75,6 @@
           android:hint="@string/login_activity_secret_hint"
           android:inputType="textVisiblePassword"
           />
-      <CheckBox
-          android:id="@+id/allcerts_checkbox"
-          android:text="@string/pref_accept_invalid_certs_summary"
-          android:layout_height="wrap_content"
-          android:layout_width="match_parent"
-          android:defaultValue="false"
-          />
       <TextView
           android:id="@+id/message_bottom"
           android:layout_height="wrap_content"
