+++ src/org/emergent/android/weave/syncadapter/AuthenticatorActivity.java
@@ -1,19 +1,3 @@
-/*
- * Copyright 2010 Patrick Woodworth
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.emergent.android.weave.syncadapter;
 
 import android.accounts.Account;
@@ -92,7 +76,7 @@ public class AuthenticatorActivity extends AccountAuthenticatorActivity
   private EditText m_passwordEdit;
   private EditText m_usernameEdit;
   private EditText m_secretEdit;
-  private CheckBox m_allcertsCheck;
+//  private CheckBox m_allcertsCheck;
 
   private int m_spinnerPos;
 
@@ -126,7 +110,7 @@ public class AuthenticatorActivity extends AccountAuthenticatorActivity
     m_usernameEdit = (EditText)findViewById(R.id.username_edit);
     m_passwordEdit = (EditText)findViewById(R.id.password_edit);
     m_secretEdit = (EditText)findViewById(R.id.secret_edit);
-    m_allcertsCheck = (CheckBox)findViewById(R.id.allcerts_checkbox);
+//    m_allcertsCheck = (CheckBox)findViewById(R.id.allcerts_checkbox);
 
     initWidgetsFromIntent(intent);
     m_message.setText(getMessage());
@@ -368,7 +352,7 @@ public class AuthenticatorActivity extends AccountAuthenticatorActivity
     initWidgetsFromIntent(this, R.id.username_edit, EXTRA_USERNAME, intent);
     initWidgetsFromIntent(this, R.id.password_edit, EXTRA_PASSWORD, intent);
     initWidgetsFromIntent(this, R.id.secret_edit, EXTRA_SECRET, intent);
-    initWidgetsFromIntent(this, R.id.allcerts_checkbox, EXTRA_ALLCERTS, intent);
+//    initWidgetsFromIntent(this, R.id.allcerts_checkbox, EXTRA_ALLCERTS, intent);
     updateSpinner();
   }
 
@@ -477,7 +461,7 @@ public class AuthenticatorActivity extends AccountAuthenticatorActivity
     String username = m_usernameEdit.getText().toString();
     String password = m_passwordEdit.getText().toString();
     String secret = m_secretEdit.getText().toString();
-    boolean allcerts = m_allcertsCheck.isChecked();
+    boolean allcerts = true; // m_allcertsCheck.isChecked();
     Log.w(TAG, String.format("Authenticating: %s : %s @ %s", username, password, serverUri));
     WeaveAccountInfo retval = createWeaveInfo(serverUri, username, password, secret);
     final SharedPreferences prefs = DobbyUtil.getAccountPreferences(AuthenticatorActivity.this, retval);
@@ -503,23 +487,6 @@ public class AuthenticatorActivity extends AccountAuthenticatorActivity
       return null;
     }
     return retval;
-  }
-
-  private static class AuthResult {
-
-    public final boolean m_success;
-    public final WeaveAccountInfo m_info;
-    public final String m_message;
-    public final Throwable m_thrown;
-
-    public AuthResult(boolean success, WeaveAccountInfo info, String message, Throwable thrown) {
-      m_success = success;
-      m_info = info;
-      m_message = message;
-      m_thrown = thrown;
-      if (m_success && (m_info == null))
-        Log.w(TAG, "AuthenticatorActivity$AuthResult.AuthResult() : success is true but info is null");
-    }
   }
 }
